Fix statgrab bug when there are no users.
I'd not added 1 to the length of the list on the basis that every name
had a space on the end, and I'd actually want to overwrite the last
one with a NULL to terminate it. But I'd not accounted for the case
when there are no users... mallocing 0 length doesn't give something
you can put a NULL in!