Bump version in preparation for new release.
We've decided to go to 0.90 because there are a number of changes to
the API that will require programs to be changed. We've put changes in
that aren't strictly necessary yet to get all the currently planned
API changes done in one go.

Also bump the shared library version accordingly.