For consistent behaviour add config.{guess,sub}.
This makes it consistent across build platforms, and ensures that
anyone checking out our Git repository gets the same versions of these
files as we do. They match upstream so can be updated when needed.

It'd be nice to be able to rely on the operating system provided ones
being current, but at least in FreeBSD they're replaced by old ones. I
consider that a bug, but it's not been resolved yet.

They're sourced from here:

http://git.savannah.gnu.org/gitweb/?p=config.git;a=blob_plain;f=config.guess;hb=HEAD
http://git.savannah.gnu.org/gitweb/?p=config.git;a=blob_plain;f=config.sub;hb=HEAD